Add enableFractionalFontSizeAndroid feature flag to render text at unrounded pixel sizes
Android rounded every text size up to a whole pixel before rendering, so
on non-integer-density devices (or with a non-default accessibility font
scale) text renders slightly too large and multi-line blocks drift
downward by up to one pixel per line.

Behind the new, default-off enableFractionalFontSizeAndroid flag:

- TextAttributeProps.fontSize and TextAttributes.effectiveFontSize keep
  the unrounded fontSize x fontScale x density value (now Float).
- ReactAbsoluteSizeSpan is a MetricAffectingSpan carrying a Float size.
- The adjustsFontSizeToFit loop scales fractional sizes from their
  original values instead of chaining truncated ratios.
- ReactTextView.setFontSize and ReactEditText pass the same float, so the
  ellipsis workaround stays consistent with the spans.
- CustomLineHeightSpan rounds to nearest instead of up.

With the flag off, every rounding step is unchanged.

+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/textinput/ReactEditText.kt
@@ -1133,7 +1133,7 @@ public open class ReactEditText public constructor(context: Context) : AppCompat
     // property hasn't been set.
     // `getEffectiveFontSize` always returns a value so don't need to check for anything like
     // `Float.NaN`.
-    setTextSize(TypedValue.COMPLEX_UNIT_PX, textAttributes.effectiveFontSize.toFloat())
+    setTextSize(TypedValue.COMPLEX_UNIT_PX, textAttributes.effectiveFontSize)
 
     val effectiveLetterSpacing = textAttributes.effectiveLetterSpacing
     if (!effectiveLetterSpacing.isNaN()) {
